requirement.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.2
djangorestframework==3.14.0
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
tzdata==2023.3
